Add delhi ndtw mesh and images of NDTW
* removed standard mesh without NDTW

In the future, the better way to manage this will be to create a pull
request for the non-NDTW mesh that I can clone. After I have it cloned,
the PR can be abandonned. I can then commit the NDTW mesh. This will
avoid adding the non-NDTW mesh to the repository.